+++ packages/frontend/src/components/project/Section.tsx
@@ -1,7 +1,7 @@
 import cx from 'classnames'
 import React, { ReactNode } from 'react'
 
-import { Heading } from '../../../components'
+import { Heading } from '../Heading'
 
 interface Props {
   title: string
